+++ lmdeploy/turbomind/deploy/source_model/__init__.py
@@ -1,8 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .baichuan import Baichuan2Model, BaichuanModel  # noqa: F401
 from .baichuan_awq import Baichuan2AwqModel, BaichuanAwqModel  # noqa: F401
-from .hf import HfModel  # noqa: F401
-from .hf_awq import HfAwqModel  # noqa: F401
 from .llama import LlamaModel  # noqa: F401
+from .llama_awq import LlamaAwqModel  # noqa: F401
+from .meta_llama import MetaLlamaModel  # noqa: F401
 from .qwen import QwenModel  # noqa: F401
 from .qwen_awq import QwenAwqModel  # noqa: F401
