+++ lmdeploy/turbomind/deploy/source_model/llama_awq.py
@@ -2,7 +2,7 @@
 import torch
 
 from .base import INPUT_MODELS
-from .hf import HfModel, HfReader
+from .llama import LlamaModel, LlamaReader
 
 
 def ensure_fp16orint32(tensors: torch.Tensor):
@@ -20,8 +20,8 @@ def ensure_fp16orint32(tensors: torch.Tensor):
     return (*result, )
 
 
-class HfAwqReader(HfReader):
-    """HfAwqReader."""
+class LlamaAwqReader(LlamaReader):
+    """LlamaAwqReader."""
 
     def __init__(self, new_params: dict, unused_params: dict, last_bin: bool):
         super().__init__(new_params, unused_params, last_bin)
@@ -52,10 +52,10 @@ class HfAwqReader(HfReader):
 
 
 @INPUT_MODELS.register_module(name='hf-awq')
-class HfAwqModel(HfModel):
-    """Awq model in hf format."""
+class LlamaAwqModel(LlamaModel):
+    """Llama Awq model in hf format."""
 
-    Reader = HfAwqReader
+    Reader = LlamaAwqReader
 
     def __init__(self,
                  model_path: str,
